mem file: change loop variable type to match numChunks type
numChunks was changed from int64_t to a size_t in a previous
commit.

+++ mxf/mxf_memory_file.c
@@ -139,7 +139,7 @@ static void mem_file_close(MXFFileSysData *sysData)
         return;
 
     if (!sysData->readOnly) {
-        int64_t i;
+        size_t i;
         for (i = 0; i < sysData->numChunks; i++)
             free(sysData->chunks[i].data);
     }
